+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - none
 
 # v2023.6.1-beta
+- sqlmath - Add sql-function vec_win_slr_updatelast().
 - sqlmath - Add sql-function vec_win_slr().
 - sqlmath - Revamp struct Vector99.
 - sqlmath - Add sql-function vec_concat().